Portfolio2-G77 3/backend/requirements.txt
mysql-connector-python
Flask
requests
Api
Flask-RESTful
